+++ src/utils/validation.ts
@@ -1,3 +1,5 @@
+import * as vscode from 'vscode';
+
 export function isValidName(name: string): boolean {
     const forbidden = /[\\\/:\*\?"<>\|]/;
     const reservedNames = [
@@ -11,5 +13,19 @@ export function isValidName(name: string): boolean {
     if (name.endsWith(' ') || name.endsWith('.')) { return false; }
     if (reservedNames.includes(name.toUpperCase())) { return false; }
 
+    return true;
+}
+
+export async function validateStructures(structures: any[]): Promise<boolean> {
+    console.log(structures);
+    if (structures.length === 0) {
+        await vscode.window.showErrorMessage('You haven\'t created any structures.\nPlease create a structure in your settings.json', { modal: true });
+        return false;
+    }
+    const emptyNameStructures = structures.filter(s => !s.name || s.name.trim() === '');
+    if (emptyNameStructures.length > 0) {
+        await vscode.window.showErrorMessage('One or more structures have an empty name. Please update your settings.', { modal: true });
+        return false;
+    }
     return true;
 }